requirements.txt
flask
lark-oapi
openai
python-dotenv
pytest
